Added support for a shutdownEnabled pin to Arduino.
This will allow us override the shutdown command normally used once watering has been skipped/performed, which is useful for service.

+++ app/src/SingleRunHandler.cpp
@@ -45,7 +45,7 @@ bool SingleRunHandler::exec() {
     break;
   }
 
-  if (!keepRunning) {
+  if (!keepRunning && arduino.isShutdownEnabled()) {
     arduino.shutdown();
     system("shutdown -h now");
   }
